backend/requirements.txt
# requirements.txt

blinker==1.8.2
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
filelock==3.15.4
Flask==3.0.3
Flask-Cors==4.0.1
future==1.0.0
gevent==24.2.1
greenlet==3.0.3
idna==3.7
importlib_metadata==8.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mysql-connector==2.2.9
pycparser==2.22
PyMySQL==1.1.1
PySocks==1.7.1
requests==2.32.3
requests-file==2.1.0
tldextract==5.1.2
urllib3==2.2.2
Werkzeug==3.0.3
zipp==3.19.2
zope.event==5.0
zope.interface==6.4.post2
